requirements.txt

aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.1.0
autobahn==24.4.2
Automat==24.8.1
awsebcli==3.20.10
beautifulsoup4==4.12.2
bleach==6.1.0
boto3==1.20.26
botocore==1.23.54
cachetools==5.3.3
cassandra-driver==3.29.1
cement==2.8.2
certifi==2023.5.7
cffi==1.17.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2024.1
crispy-bootstrap5==0.7
cryptography==41.0.2
daphne==4.0.0
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==0.5.0
Django==4.2
django-admin-sortable2==2.1.10
django-anymail==9.1
django-bootstrap4==23.2
django-ckeditor==6.7.0
django-ckeditor-5==0.2.10
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-dbbackup==4.0.2
django-dotenv==1.4.2
django-environ==0.9.0
django-extensions==3.2.3
django-import-export==3.2.0
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-phonenumber-field==8.0.0
django-recaptcha==3.0.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-six==1.0.5
django-static-fontawesome==5.14.0.0
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
environs==9.5.0
et-xmlfile==1.1.0
frozenlist==1.4.1
geoip2==4.6.0
geomet==0.2.1.post1
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gunicorn==21.2.0
httplib2==0.22.0
humanize==4.6.0
hyperlink==21.0.0
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.4.9
incremental==24.7.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
Markdown==3.4.3
MarkupPy==1.14
MarkupSafe==2.1.3
marshmallow==3.19.0
maxminddb==2.4.0
moviepy==1.0.3
msal==1.30.0
msgpack==1.0.7
multidict==6.1.0
numpy==1.26.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.5
packaging==23.2
pandas==2.1.1
pathspec==0.10.1
phonenumbers==8.13.47
pillow==10.4.0
proglog==0.1.10
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pycparser==2.22
PyJWT==2.6.0
pyOpenSSL==23.2.0
pyparsing==3.1.4
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
razorpay==1.4.2
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.1.0
rsa==4.9
s3transfer==0.5.2
semantic-version==2.8.5
service-identity==23.1.0
setuptools==75.1.0
shortuuid==1.0.11
six==1.16.0
soupsieve==2.6
sqlparse==0.4.4
stripe==7.0.0
tablib==3.6.1
termcolor==1.1.0
tqdm==4.66.1
twilio==8.5.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2023.3
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.17
user-agents==2.2.0
wcwidth==0.1.9
webencodings==0.5.1
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.13.0
zope.interface==7.0.3
